#328: Moved the materia models to model/items/materia
-
