+++ src/app/ffbe/mappers/materia-mapper.spec.ts
@@ -1,7 +1,7 @@
 import {Objet} from '../model/objet/objet.model';
 import {MateriaMapper} from './materia-mapper';
-import {MATERIAS_TEST_DATA} from '../model/materia.model.spec';
-import {Materia} from '../model/materia.model';
+import {MATERIAS_TEST_DATA} from '../model/items/materia/materia.model.spec';
+import {Materia} from '../model/items/materia/materia.model';
 import {FfbeUtils} from '../utils/ffbe-utils';
 import {PASSIVE_SKILLS_TEST_DATA} from '../model/skill.model.spec';
 import {Skill} from '../model/skill.model';
