+++ src/app/ffbe/mappers/skill-mapper.spec.ts
@@ -14,10 +14,10 @@ import {EQUIPMENTS_TEST_DATA} from '../model/items/equipment/equipment.model.spe
 import {Equipment} from '../model/items/equipment/equipment.model';
 import {EquipmentsServiceMock} from '../services/equipments.service.spec';
 import {EquipmentsService} from '../services/equipments.service';
-import {MATERIAS_TEST_DATA} from '../model/materia.model.spec';
+import {MATERIAS_TEST_DATA} from '../model/items/materia/materia.model.spec';
 import {MateriasService} from '../services/materias.service';
 import {MateriasServiceMock} from '../services/materias.service.spec';
-import {Materia} from '../model/materia.model';
+import {Materia} from '../model/items/materia/materia.model';
 
 describe('SkillMapper', () => {
   it('should transform ability icon string to number', () => {
