+++ src/app/ffbe/model/items/materia/materia-strings.model.ts
